Deleted sqllogictest code that was moved to its own repo. Ran go mod tidy
