+++ go/libraries/doltcore/sqle/logictest/main/main.go
@@ -12,9 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package logictest contains code to define a harness to execute sqllogictests and a runner to run them, as well as a
-// parser to parse the result log of such a test run.
-// https://www.sqlite.org/sqllogictest/doc/trunk/about.wiki
-// A git-hub version of the original sqllogictest with all test cases can be found here:
-// https://github.com/gregrahn/sqllogictest
-package logictest
+package main
+
+import (
+	"os"
+
+	"github.com/liquidata-inc/sqllogictest/go/logictest"
+
+	"github.com/liquidata-inc/dolt/go/libraries/doltcore/sqle/logictest/dolt"
+)
+
+// Runs all sqllogictest test files (or directories containing them) given as arguments.
+func main() {
+	args := os.Args[1:]
+	h := &dolt.DoltHarness{}
+	logictest.RunTestFiles(h, args...)
+}
