--- go/libraries/doltcore/sqle/logictest/parser/linescanner.go
@@ -1,28 +1,0 @@
-// Copyright 2019 Liquidata, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package parser
-
-import "bufio"
-
-// LineScanner wraps a standard bufio.Scanner and adds line numbers for debugging / reporting.
-type LineScanner struct {
-	*bufio.Scanner
-	LineNum int
-}
-
-func (ls *LineScanner) Scan() bool {
-	ls.LineNum++
-	return ls.Scanner.Scan()
-}
