--- go/libraries/doltcore/sqle/logictest/parser/testdata/select1.test
@@ -1,77 +1,0 @@
-statement ok
-CREATE TABLE t1(a INTEGER, b INTEGER, c INTEGER, d INTEGER, e INTEGER)
-
-statement ok
-INSERT INTO t1(e,c,b,d,a) VALUES(103,102,100,101,104)
-
-statement error # a comment about this error
-INSERT INTO t1(a,c,d,e,b) VALUES(107,106,108,109,105)
-
-# you wouldn't see this in a real test file, but naked halts can be inserted for debugging
-halt
-
-query I nosort
-SELECT CASE WHEN c>(SELECT avg(c) FROM t1) THEN a*2 ELSE b*10 END
-  FROM t1
- ORDER BY 1
-----
-30 values hashing to 3c13dee48d9356ae19af2515e05e6b54
-
-hash-threshold 8
-
-# A long comment
-# That goes on multiple lines
-
-# Followed by another one
-# With a blank line in between
-
-query II nosort label-1
-SELECT a+b*2+c*3+d*4+e*5,
-       (a+b+c+d+e)/5
-  FROM t1
- ORDER BY 1,2
-----
-60 values hashing to 808146289313018fce25f1a280bd8c30
-
-onlyif mysql # tests below here won't work on mysql
-halt
-
-onlyif mysql # Comment
-query IIIII rowsort
-SELECT a+b*2+c*3+d*4+e*5,
-       CASE WHEN a<b-3 THEN 111 WHEN a<=b THEN 222
-        WHEN a<b+3 THEN 333 ELSE 444 END,
-       abs(b-c),
-       (a+b+c+d+e)/5,
-       a+b*2+c*3
-  FROM t1
- WHERE (e>c OR e<d)
-   AND d>e
-   AND EXISTS(SELECT 1 FROM t1 AS x WHERE x.b<t1.b)
- ORDER BY 4,2,1,3,5
-----
-1
-2
-3
-4
-5
-
-# This is a comment
-skipif mssql
-query II nosort  # This is a comment
-SELECT a-b,
-       CASE WHEN a<b-3 THEN 111 WHEN a<=b THEN 222
-        WHEN a<b+3 THEN 333 ELSE 444 END
-  FROM t1
- WHERE c>d
-   AND b>c
- ORDER BY 2,1
-----
--3
-222
--3
-222
--1
-222
--1
-222
