--- go/libraries/doltcore/sqle/logictest/testdata/resultlog.txt
@@ -1,10 +1,0 @@
-2019-10-16T16:02:18.3408696-07:00 evidence/in1.test:25: SELECT 1 IN () skipped
-2019-10-16T16:02:18.3418683-07:00 evidence/in1.test:30: SELECT 1 IN (2) ok
-2019-10-16T16:02:18.3418683-07:00 evidence/in1.test:35: SELECT 1 IN (2,3,4,5,6,7,8,9) ok
-2019-10-16T16:02:18.3418683-07:00 evidence/in1.test:41: SELECT 1 NOT IN () skipped
-2019-10-16T16:02:18.3418683-07:00 evidence/in1.test:46: SELECT 1 NOT IN (2) ok
-2019-10-16T16:02:18.3418683-07:00 evidence/in1.test:51: SELECT 1 NOT IN (2,3,4,5,6,7,8,9) ok
-2019-10-16T16:02:18.3418683-07:00 evidence/in1.test:57: SELECT null IN () skipped
-2019-10-16T16:02:18.3418683-07:00 evidence/in1.test:63: SELECT null NOT IN () skipped
-2019-10-16T16:02:18.3428692-07:00 evidence/in1.test:68: CREATE TABLE t1(x INTEGER) not ok: Unexpected error no primary key columns
-2019-10-16T16:02:18.3428692-07:00 evidence/in1.test:72: SELECT 1 IN t1 skipped
